Core Features implemented with savable favorites in research and power. Some bugs still exists that needs some hammering on...

+++ src/main/java/hok/chompzki/biocristals/blocks/croot/ICrootPowerCon.java
@@ -2,7 +2,7 @@ package hok.chompzki.biocristals.blocks.croot;
 
 public interface ICrootPowerCon {
 	
-	public float getConsumption();
+	public int getConsumption();
 	
 	public int getX();
